+++ AST/Record.h
@@ -52,7 +52,7 @@ class Record : public Expression, public HasScope
 public:
 	Record(UniqPtr<Scope>& fields, const RecordType&, const SourceRange&);
 
-	virtual void PrettyPrint(Bytestream&, size_t indent = 0) const override;
+	virtual void PrettyPrint(Bytestream&, unsigned int indent = 0) const override;
 	virtual void Accept(Visitor&) const override;
 
 	virtual dag::ValuePtr evaluate(EvalContext&) const override;
