+++ AST/literals.h
@@ -72,7 +72,7 @@ public:
 	std::string str() const override;
 
 	virtual dag::ValuePtr evaluate(EvalContext&) const override;
-	virtual void PrettyPrint(Bytestream&, size_t indent = 0) const override;
+	virtual void PrettyPrint(Bytestream&, unsigned int indent = 0) const override;
 	virtual void Accept(Visitor&) const override;
 };
 
@@ -89,7 +89,7 @@ public:
 	std::string str() const override;
 
 	virtual dag::ValuePtr evaluate(EvalContext&) const override;
-	virtual void PrettyPrint(Bytestream&, size_t indent = 0) const override;
+	virtual void PrettyPrint(Bytestream&, unsigned int indent = 0) const override;
 	virtual void Accept(Visitor&) const override;
 };
 
@@ -107,7 +107,7 @@ public:
 	std::string str() const override;
 
 	virtual dag::ValuePtr evaluate(EvalContext&) const override;
-	virtual void PrettyPrint(Bytestream&, size_t indent = 0) const override;
+	virtual void PrettyPrint(Bytestream&, unsigned int indent = 0) const override;
 	virtual void Accept(Visitor&) const override;
 };
